feat: 소켓을 리덕스 미들웨어에서 관리
* 리덕스 미들웨어에서 소켓을 관리하면 모든 페이지에서 연결을 끊지 않고 소켓 클라이언트를 공유할 수 있다.
* 현재 로그인/내 정보 요청을 성공했을 때 소켓을 연결하도록 해두었고, 로그아웃 시 소켓의 연결이 끊긴다.
